create fake webpage
create the webpage with fake data, and without connecting to a server.
it was mostly to see how it will look like but also understand what to do and create the needed css.
a lot of the stuff in the html will be changes in server (like almost everything) and I also need to create the refresh and GET functions. so still have work but those are the fake stuff. and I don't really think I need comments here.