create the decoder
the decoder right now it's for the sql only and is used to take raw hex and make it a list for the sql then to interpret.
comments I will add in the next upload but we also have in the example[...].py the way to add it to the server tables.
what needed to be done is now to take the data from satnogs, map it and add it.
when it will work I need to make it to the html page and then https webserver.
also I think i will make the config to be read at start and then have a dict for everything but also without doubles.
I have a lot more stuff to do so hopefully I will make it work.

ALSO IMPORTANT right now I don't have edge check and I need to do some and write them.