the website works!
I have the base of the website (still need to do the scroll up and down). and also open the thread of satNogs.

need to add a lot of comments and error catching but the basic of it is finished.

this program need to run from the start. but if not we can get most of the first packets as a download and other.

still need to create format and the csv download but I think I can do it fast.
also need to understand why the favicon is not passing.